chore: Mockup of updated layout of bottom actions
